// client/src/models/solicits-with-address.ts
export interface SolicitsWithAddress {
	_id: string;
	product_name: string;
	list_id: string; 
	product_cost?: Number;
	cost_unit?: string;
	description?: string;
	still_active?: boolean;
	still_available?: boolean;
	food_pic?: string;
	seller_sub?: string;
	seller_nickname?: string;
	product_id?: string;
    address_id?: string;
    coordinates: Number[];
    
}